+++ lib/src/bones_api_entity_db_mysql.dart
@@ -767,18 +767,13 @@ class DBMySQLAdapter extends DBSQLAdapter<DBMySqlConnectionWrapper>
               _DBMySqlConnectionTransaction(connection.connection, t);
           contextCompleter.complete(transactionWrap);
 
-          return transaction.transactionFuture.then((ret) {
-            // When aborted `_transactionCompleter.complete` will be called
-            // with the error (not calling `completeError`), since it's
-            // running in another error zone (won't reach `onError`):
-            if (ret is TransactionAbortedError) {
-              throw ret;
-            }
-            return ret;
-          }, onError: (e, s) {
-            cancelTransaction(transaction, connection, e, s);
-            throw e;
-          });
+          return transaction.transactionFuture.then(
+            (res) => resolveTransactionResult(res, transaction, connection),
+            onError: (e, s) {
+              cancelTransaction(transaction, connection, e, s);
+              throw e;
+            },
+          );
         });
       },
       validator: (c) => !transaction.isAborted,
@@ -794,6 +789,12 @@ class DBMySQLAdapter extends DBSQLAdapter<DBMySqlConnectionWrapper>
 
     return contextCompleter.future;
   }
+
+  @override
+  bool get cancelTransactionResultWithError => false;
+
+  @override
+  bool get throwTransactionResultWithError => true;
 
   @override
   bool cancelTransaction(