+++ test/bones_api_entity_db_tests_base.dart
@@ -426,14 +426,17 @@ Future<bool> runAdapterTests(
     });
 
     test('TestEntityRepositoryProvider', () async {
-      var storeAPIRepository = entityRepositoryProvider.storeAPIRepository;
-      var addressAPIRepository = entityRepositoryProvider.addressAPIRepository;
-      var roleAPIRepository = entityRepositoryProvider.roleAPIRepository;
-      var userInfoAPIRepository =
+      final sqlAdapter = await entityRepositoryProvider.adapter;
+      final storeAPIRepository = entityRepositoryProvider.storeAPIRepository;
+      final addressAPIRepository =
+          entityRepositoryProvider.addressAPIRepository;
+      final roleAPIRepository = entityRepositoryProvider.roleAPIRepository;
+      final userInfoAPIRepository =
           entityRepositoryProvider.userInfoAPIRepository;
-      var userAPIRepository = entityRepositoryProvider.userAPIRepository;
-      var photoAPIRepository = entityRepositoryProvider2.photoAPIRepository;
+      final userAPIRepository = entityRepositoryProvider.userAPIRepository;
+      final photoAPIRepository = entityRepositoryProvider2.photoAPIRepository;
 
+      expect(sqlAdapter.isInitialized, isTrue);
       expect(await userAPIRepository.length(), equals(0));
 
       {
@@ -707,20 +710,57 @@ Future<bool> runAdapterTests(
         var userDuplicated = User('joe@$testDomain', '456', address, [role],
             level: 100, creationTime: user1CreationTime);
 
-        EntityFieldInvalid? error;
-        try {
-          await userAPIRepository.store(userDuplicated);
-        } on EntityFieldInvalid catch (e) {
-          error = e;
+        {
+          EntityFieldInvalid? error;
+          int? result;
+          try {
+            result = await userAPIRepository.store(userDuplicated);
+          } on EntityFieldInvalid catch (e) {
+            error = e;
+          }
+
+          expect(result, isNull);
+
+          expect(error, isA<EntityFieldInvalid>());
+          expect(
+            error.toString(),
+            matches(RegExp(
+                r'Invalid entity\((?:User)?@table:user\) field(?:\(.*?email.*?\))?> reason: unique ; value: <.*?joe@[\w.+]+\.com.*?>.*',
+                dotAll: true)),
+          );
         }
 
-        expect(error, isA<EntityFieldInvalid>());
-        expect(
-          error.toString(),
-          matches(RegExp(
-              r'Invalid entity\((?:User)?@table:user\) field(?:\(.*?email.*?\))?> reason: unique ; value: <.*?joe@[\w.+]+\.com.*?>.*',
-              dotAll: true)),
-        );
+        if (sqlAdapter.capability.transactionAbort) {
+          var transaction = Transaction();
+
+          var result = await transaction.execute(() async {
+            var prevUsers = await userAPIRepository
+                .selectFirstByQuery('email == "${userDuplicated.email}"');
+            expect(prevUsers, isNotNull);
+            expect(prevUsers?.email, equals(userDuplicated.email));
+            var id = await userAPIRepository.store(userDuplicated);
+            return id as int?;
+          });
+
+          print(transaction);
+
+          expect(result, isNull);
+
+          expect(transaction.isOpen, isTrue);
+          expect(transaction.isAborted, isTrue);
+          expect(transaction.isCommitted, isFalse);
+          expect(transaction.length, greaterThanOrEqualTo(1));
+          expect(transaction.abortError, isNotNull);
+          expect(transaction.abortError?.reason, isNull);
+
+          expect(transaction.abortError?.error, isA<EntityFieldInvalid>());
+          expect(
+            transaction.abortError?.error.toString(),
+            matches(RegExp(
+                r'Invalid entity\((?:User)?@table:user\) field(?:\(.*?email.*?\))?> reason: unique ; value: <.*?joe@[\w.+]+\.com.*?>.*',
+                dotAll: true)),
+          );
+        }
       }
 
       var user2CreationTime = DateTime.utc(2021, 9, 21, 22, 11, 12, 0, 0);
@@ -1392,10 +1432,8 @@ Future<bool> runAdapterTests(
         expect(transaction.isAborted, isFalse);
         expect(transaction.isCommitted, isTrue);
         expect(transaction.length, greaterThanOrEqualTo(7));
-        expect(transaction.abortedError, isNull);
+        expect(transaction.abortError, isNull);
       }
-
-      var sqlAdapter = await entityRepositoryProvider.adapter;
 
       // If `Transaction.abort` is supported:
       if (sqlAdapter.capability.transactionAbort) {
@@ -1427,8 +1465,8 @@ Future<bool> runAdapterTests(
         expect(transaction.isAborted, isTrue);
         expect(transaction.isCommitted, isFalse);
         expect(transaction.length, greaterThanOrEqualTo(6));
-        expect(transaction.abortedError, isNotNull);
-        expect(transaction.abortedError?.reason, equals('Test'));
+        expect(transaction.abortError, isNotNull);
+        expect(transaction.abortError?.reason, equals('Test'));
       }
 
       {